requirements.txt
Django>=1.8.1,<1.8.99
dj-static
mysqlclient
django-reversion
markdown
django-storages-redux
boto
urllib3